requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

langchain-google-genai>=2.0.0  
langchain-anthropic>=0.3.0   
langchain-openai>=0.2.0        

chromadb>=0.5.0
langchain-chroma>=0.1.0

tavily-python>=0.3.0             # Set TAVILY_API_KEY in .env

aiohttp>=3.9.0
tqdm>=4.66.0

streamlit>=1.38.0
plotly>=5.22.0
pandas>=2.2.0

python-dotenv>=1.0.0